# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "University API",
        "description": "University API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://192.168.0.109:8000"
        }
    ],
    "paths": {
        "/api/faculties": {
            "get": {
                "tags": [
                    "faculties"
                ],
                "summary": "Get list of faculties",
                "description": "Returns list of faculties",
                "operationId": "getFacultiesList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "faculties"
                ],
                "summary": "Endpoint for creating a new faculty.",
                "description": "Create faculty",
                "operationId": "createFaculty",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/faculties/{id}": {
            "get": {
                "tags": [
                    "faculties"
                ],
                "summary": "Get single faculty",
                "description": "Returns single faculty",
                "operationId": "getSingleFaculty",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "faculties"
                ],
                "summary": "Update single faculty",
                "description": "Endpoint for update single faculty",
                "operationId": "UpdateSingleFaculty",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "faculties"
                ],
                "summary": "Endpoint for deleting faculty.",
                "description": "Delete faculty",
                "operationId": "deleteFaculty",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/faculties/{id}/groups": {
            "get": {
                "tags": [
                    "faculties"
                ],
                "summary": "Get list of groups of faculty",
                "description": "Returns list of groups of faculty",
                "operationId": "getGroupsListOfFaculty",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
    }
}